Rename scenario `base` to `basic`
