+++ webhosting-operator/pkg/experiment/scenario/all/all.go
@@ -18,6 +18,6 @@ limitations under the License.
 package all
 
 import (
-	_ "github.com/timebertt/kubernetes-controller-sharding/webhosting-operator/pkg/experiment/scenario/base"
+	_ "github.com/timebertt/kubernetes-controller-sharding/webhosting-operator/pkg/experiment/scenario/basic"
 	_ "github.com/timebertt/kubernetes-controller-sharding/webhosting-operator/pkg/experiment/scenario/reconcile"
 )
